Move header under main page
