+++ web/src/app/components/main-page/main-page.module.ts
@@ -26,7 +26,7 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { NgModule } from '@angular/core';
 import { SidePanelModule } from 'app/components/side-panel/side-panel.module';
-import { SurveyHeaderModule } from 'app/components/survey-header/survey-header.module';
+import { SurveyHeaderModule } from 'app/components/main-page/survey-header/survey-header.module';
 import { MatIconModule } from '@angular/material/icon';
 
 @NgModule({
