+++ web/src/app/components/main-page/survey-header/survey-header.component.spec.ts
